split prettier plugin into multiple files
